Move test to proper location
I'd overlooked that the `state_machine.api` items are re-exported
by the `state_machine` module. D'Oh.
